Workaround for negative horiBearingX
Many letters is my default latin font contain glyphs with negative
horiBearingX. Since it's cast to usize it ends up as an enormous
positive value moving the glyph out of the canvas. What I'm proposing
here may not be a proper solution, but most of the glyphs seems to be
placed properly.